issue #3: empty query exception
