+++ tests/test_azure_search.py
@@ -1,7 +1,11 @@
 import unittest
 from unittest.mock import patch
 
-from app.azure_search import AzureSearchQueryError, search_documents
+from app.azure_search import (
+    AzureSearchQueryError,
+    EmptyQueryError,
+    azure_search_documents,
+)
 from tests.common import TestAzureSearchConfig
 
 
@@ -10,14 +14,14 @@ class TestAzureSearch(unittest.TestCase):
         self.config = TestAzureSearchConfig()
 
     def test_search_documents_empty_query(self):
-        results = search_documents("", self.config)
-        self.assertEqual(results, [])
+        with self.assertRaises(EmptyQueryError):
+            azure_search_documents("", self.config)
 
     @patch("app.azure_search.logging")
     def test_search_documents_query_error(self, mock_logging):
         self.config.client.search.side_effect = Exception("Search failed")
         with self.assertRaises(AzureSearchQueryError):
-            search_documents("test_query", self.config)
+            azure_search_documents("test_query", self.config)
         mock_logging.error.assert_called()
