--- todo.md
@@ -1,18 +1,0 @@
-# Azure Cognitive Search Integration
-
-- Set Up Azure Cognitive Search Connection
-  - [X] Prepare the development environment.
-  - [X] Establish a secure connection to the Azure Cognitive Search service.
-  - [X] Test connection.
-
-- Backend API Endpoint Creation
-  - [X] Implement a new API endpoint for handling search queries.
-  - [X] Manual test the endpoint.
-
-- Testing and Debugging
-  - [X] Write unit and integration tests for the new search functionality.
-  - [ ] Test the endpoint with different search scenarios and document any bugs.
-
-- Documentation
-  - [ ] Document the setup process, API usage, and configuration details.
-  - [ ] Update the project README with the new feature information.
